feat: embed native black box solver
