+++ acvm_js/Cargo.toml
@@ -17,7 +17,7 @@ crate-type = ["cdylib"]
 cfg-if = "1.0.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-acvm = { version = "0.21.0", path = "../acvm", default-features = false }
+acvm = { version = "0.21.0", path = "../acvm", default-features = false, features = ["wasm"] }
 wasm-bindgen = { version = "0.2.87", features = ["serde-serialize"] }
 wasm-bindgen-futures = "0.4.36"
 serde = { version = "1.0.136", features = ["derive"] }
