+++ blackbox_solver/Cargo.toml
@@ -32,29 +32,35 @@ p256 = { version = "0.11.0", features = [
     "arithmetic",
 ] }
 
+
+# Native
+barretenberg-sys = { version = "0.2.0", optional = true }
+
 # Barretenberg WASM dependencies
-rust-embed = { version = "6.6.0", features = [
+rust-embed = { version = "6.6.0", optional = true, features = [
     "debug-embed",
     "interpolate-folder-path",
     "include-exclude",
 ] }
 
+# WASM imports differ based on target
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-wasmer = { version = "3.3", default-features = false, features = [ "js-default" ] }
-getrandom = { version = "0.2", features = [ "js" ]}
-wasm-bindgen-futures = "0.4.36"
-js-sys = "0.3.62"
-
+wasmer = { version = "3.3", default-features = false, features = [ "js-default" ], optional = true }
+getrandom = { version = "0.2", features = [ "js" ], optional = true}
+wasm-bindgen-futures = { version = "0.4.36" , optional = true }
+js-sys = { version = "0.3.62" , optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-getrandom = "0.2"
-wasmer = "3.3"
+getrandom = { version = "0.2", optional = true }
+wasmer = { version = "3.3", optional = true }
 
 
 [build-dependencies]
 pkg-config = "0.3"
 
 [features]
-default = ["bn254"]
+default = ["bn254", "wasm"]
 bn254 = ["acir/bn254"]
 bls12_381 = ["acir/bls12_381"]
+native = ["dep:barretenberg-sys"]
+wasm = ["dep:wasmer", "dep:rust-embed", "dep:getrandom"]
