+++ blackbox_solver/src/barretenberg/barretenberg_structures.rs
