+++ brillig_vm/Cargo.toml
@@ -17,6 +17,8 @@ num-bigint.workspace = true
 num-traits.workspace = true
 
 [features]
-default = ["bn254"]
+default = ["bn254", "native"]
 bn254 = ["acir/bn254"]
 bls12_381 = ["acir/bls12_381"]
+native = ["blackbox_solver/native"]
+wasm = ["blackbox_solver/wasm"]
