+++ flake.nix
@@ -88,13 +88,33 @@
         # Note: Setting this allows for consistent behavior across build and shells, but is mostly
         # hidden from the developer - i.e. when they see the command being run via `nix flake check`
         # RUST_TEST_THREADS = "1";
+      };
+
+      nativeEnvironment = sharedEnvironment // {
+        # rust-bindgen needs to know the location of libclang
+        LIBCLANG_PATH = "${pkgs.llvmPackages.libclang.lib}/lib";
+      };
+
+      wasmEnvironment = sharedEnvironment // {
         BARRETENBERG_BIN_DIR = "${pkgs.barretenberg-wasm}/bin";
       };
 
-      wasmEnvironment = sharedEnvironment // { };
-
       sourceFilter = path: type:
         (craneLib.filterCargoSources path type);
+
+      # As per https://discourse.nixos.org/t/gcc11stdenv-and-clang/17734/7 since it seems that aarch64-linux uses
+      # gcc9 instead of gcc11 for the C++ stdlib, while all other targets we support provide the correct libstdc++
+      stdenv =
+        if (pkgs.stdenv.targetPlatform.isGnu && pkgs.stdenv.targetPlatform.isAarch64) then
+          pkgs.overrideCC pkgs.llvmPackages.stdenv (pkgs.llvmPackages.clang.override { gccForLibs = pkgs.gcc11.cc; })
+        else
+          pkgs.llvmPackages.stdenv;
+
+      extraBuildInputs = pkgs.lib.optionals pkgs.stdenv.isDarwin [
+        # Need libiconv and apple Security on Darwin. See https://github.com/ipetkov/crane/issues/156
+        pkgs.libiconv
+        pkgs.darwin.apple_sdk.frameworks.Security
+      ];
 
       # The `self.rev` property is only available when the working tree is not dirty
       GIT_COMMIT = if (self ? rev) then self.rev else "unknown";
@@ -109,11 +129,36 @@
           filter = sourceFilter;
         };
 
-        cargoClippyExtraArgs = "--package acvm_js --all-targets -- -D warnings";
         cargoTestExtraArgs = "--workspace";
 
         # We don't want to run checks or tests when just building the project
         doCheck = false;
+      };
+
+      # Combine the environment and other configuration needed for crane to build with the native feature
+      nativeArgs = nativeEnvironment // commonArgs // {
+        # Use our custom stdenv to build and test our Rust project
+        inherit stdenv;
+
+        nativeBuildInputs = [
+          # This provides the pkg-config tool to find barretenberg & other native libraries
+          pkgs.pkg-config
+          # This provides the `lld` linker to cargo
+          pkgs.llvmPackages.bintools
+        ];
+
+        buildInputs = [
+          pkgs.llvmPackages.openmp
+          pkgs.barretenberg
+        ] ++ extraBuildInputs;
+      };
+
+      # Combine the environment and other configuration needed for crane to build with the native feature
+      wasmArgs = wasmEnvironment // commonArgs // {
+        # Use our custom stdenv to build and test our Rust project
+        inherit stdenv;
+
+        cargoExtraArgs = "--no-default-features --features='bn254, wasm'";
       };
 
       # Combine the environment and other configuration needed for crane to build with the wasm feature
@@ -121,7 +166,7 @@
 
         inherit (crateACVMJSDefinitions) pname version;
 
-        cargoExtraArgs = "--package acvm_js --target=wasm32-unknown-unknown";
+        cargoExtraArgs = "--package acvm_js --target=wasm32-unknown-unknown --no-default-features --features='bn254'";
 
         cargoVendorDir = craneLib.vendorCargoDeps {
           src = ./.;
@@ -131,8 +176,26 @@
 
       };
 
+      ## ACVM Rust Library
+
       # Build *just* the cargo dependencies, so we can reuse all of that work between runs
-      cargo-artifacts = craneLib.buildDepsOnly commonArgs;
+      acvm-native-cargo-artifacts = craneLib.buildDepsOnly nativeArgs;
+      acvm-wasm-cargo-artifacts = craneLib.buildDepsOnly commonArgs;
+      acvm-js-cargo-artifacts = craneLib.buildDepsOnly acvmjsWasmArgs;
+
+      acvm-native = craneLib.buildPackage (nativeArgs // {
+        inherit GIT_COMMIT GIT_DIRTY;
+
+        cargoArtifacts = acvm-native-cargo-artifacts;
+      });
+
+      acvm-wasm = craneLib.buildPackage (wasmArgs // {
+        inherit GIT_COMMIT GIT_DIRTY;
+
+        cargoArtifacts = acvm-wasm-cargo-artifacts;
+      });
+
+      ## ACVM JS stuff
 
       wasm-bindgen-cli = pkgs.callPackage ./acvm_js/nix/wasm-bindgen-cli/default.nix {
         rustPlatform = pkgs.makeRustPlatform {
@@ -141,6 +204,37 @@
         };
       };
 
+      acvm-js = craneLib.buildPackage (acvmjsWasmArgs // {
+        cargoArtifacts = acvm-js-cargo-artifacts;
+
+        inherit GIT_COMMIT;
+        inherit GIT_DIRTY;
+
+        COMMIT_SHORT = builtins.substring 0 7 GIT_COMMIT;
+        VERSION_APPENDIX = if GIT_DIRTY == "true" then "-dirty" else "";
+
+        src = ./.; #craneLib.cleanCargoSource (craneLib.path ./.);
+
+        nativeBuildInputs = with pkgs; [
+          binaryen
+          which
+          git
+          jq
+          rustToolchain
+          wasm-bindgen-cli
+        ];
+
+        CARGO_TARGET_DIR = "./target";
+
+        buildPhaseCargoCommand = ''
+          bash ./acvm_js/buildPhaseCargoCommand.sh
+        '';
+
+        installPhase = ''
+          bash ./acvm_js/installPhase.sh        
+        '';
+      });
+
     in
     rec {
       checks = {
@@ -148,60 +242,32 @@
         cargo-clippy = craneLib.cargoClippy (commonArgs // sharedEnvironment // {
           inherit GIT_COMMIT GIT_DIRTY;
 
-          cargoArtifacts = cargo-artifacts;
+          cargoArtifacts = acvm-native-cargo-artifacts;
           doCheck = true;
         });
 
         cargo-test = craneLib.cargoTest (commonArgs // sharedEnvironment // {
           inherit GIT_COMMIT GIT_DIRTY;
 
-          cargoArtifacts = cargo-artifacts;
+          cargoArtifacts = acvm-native-cargo-artifacts;
           doCheck = true;
         });
 
         cargo-fmt = craneLib.cargoFmt (commonArgs // sharedEnvironment // {
           inherit GIT_COMMIT GIT_DIRTY;
 
-          cargoArtifacts = cargo-artifacts;
+          cargoArtifacts = acvm-native-cargo-artifacts;
           doCheck = true;
         });
 
       };
 
       packages = {
-        default = craneLib.mkCargoDerivation (acvmjsWasmArgs // rec {
+        inherit acvm-native;
+        inherit acvm-wasm;
+        inherit acvm-js;
 
-          cargoArtifacts = cargo-artifacts;
-
-          inherit GIT_COMMIT;
-          inherit GIT_DIRTY;
-
-          COMMIT_SHORT = builtins.substring 0 7 GIT_COMMIT;
-          VERSION_APPENDIX = if GIT_DIRTY == "true" then "-dirty" else "";
-
-          src = ./.; #craneLib.cleanCargoSource (craneLib.path ./.);
-
-          nativeBuildInputs = with pkgs; [
-            binaryen
-            which
-            git
-            jq
-            rustToolchain
-            wasm-bindgen-cli
-          ];
-
-          CARGO_TARGET_DIR = "./target";
-
-          buildPhaseCargoCommand = ''
-            bash ./acvm_js/buildPhaseCargoCommand.sh
-          '';
-
-          installPhase = ''
-            bash ./acvm_js/installPhase.sh        
-          '';
-
-        });
-        inherit cargo-artifacts;
+        default = acvm-native;
       };
 
       # Setup the environment to match the stdenv from `nix build` & `nix flake check`, and
